% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive_diff.R
\name{consecutive_diff}
\alias{consecutive_diff}
\title{Calculate the consecutive difference for a sequence of numbers}
\usage{
consecutive_diff(number_sequence)
}
\arguments{
\item{number_sequence}{the number sequence the user chooses to put in this function. The input should be a numeric vector, or list of single numerics.}
}
\value{
consecutive_diff() returns a numeric vector
}
\description{
given a sequence such as (x1, x2, x3, x4, x5, ...),
consecutive_diff() can calculate all the consecutive differences as {(x2-x1), (x3-x2), (x4-x3), ...}
}
\examples{
# create an array of consecutive squares and calculate the consecutive differences
lst <- c(0, 1, 4, 9, 16, 25, 36, 49, 64)
consecutive_diff(lst)
# generate a series of random integers and calculate the consecutive difference
consecutive_diff(sample.int(100,20))
}
